chore(QiScans): move fetch functions to network.ts
