+++ src/QiScans/implementations/manga/main.ts
@@ -1,7 +1,7 @@
 import type { Request, SourceManga } from "@paperback/types";
 import { QISCANS_API_BASE } from "../../main";
 import type { QIScansPost } from "../shared/models";
-import { fetchJSON } from "../shared/utils";
+import { fetchJSON } from "../../services/network";
 import { parseMangaDetails } from "./parsers";
 
 export class MangaProvider {
